refactor: Move SimpleFileTransfer to an utility file
