+++ tests/flytekit/integration/remote/test_remote.py
@@ -28,7 +28,9 @@ from flyteidl.service import dataproxy_pb2 as _data_proxy_pb2
 from flytekit.types.schema import FlyteSchema
 from flytekit.clients.friendly import SynchronousFlyteClient as _SynchronousFlyteClient
 from flytekit.configuration import PlatformConfig
-from botocore.client import BaseClient
+
+from tests.flytekit.integration.remote.utils import SimpleFileTransfer
+
 
 MODULE_PATH = pathlib.Path(__file__).parent / "workflows/basic"
 CONFIG = os.environ.get("FLYTECTL_CONFIG", str(pathlib.Path.home() / ".flyte" / "config-sandbox.yaml"))
@@ -805,78 +807,6 @@ def test_get_control_plane_version():
     client = _SynchronousFlyteClient(PlatformConfig.for_endpoint("localhost:30080", True))
     version = client.get_control_plane_version()
     assert version == "unknown" or version.startswith("v")
-
-
-class SimpleFileTransfer:
-    """Utilities for file transfer to minio s3 bucket.
-
-    Mainly support single file uploading and automatic teardown.
-    """
-
-    def __init__(self) -> None:
-        self._remote = FlyteRemote(
-            config=Config.auto(config_file=CONFIG),
-            default_project=PROJECT,
-            default_domain=DOMAIN
-        )
-        self._s3_client = self._get_minio_s3_client(self._remote)
-
-    def _get_minio_s3_client(self, remote: FlyteRemote) -> BaseClient:
-        """Creat a botocore client."""
-        minio_s3_config = remote.file_access.data_config.s3
-        sess = botocore.session.get_session()
-
-        return sess.create_client(
-            "s3",
-            endpoint_url=minio_s3_config.endpoint,
-            aws_access_key_id=minio_s3_config.access_key_id,
-            aws_secret_access_key=minio_s3_config.secret_access_key,
-        )
-
-    def upload_file(self, file_type: str) -> str:
-        """Upload a single file to minio s3 bucket.
-
-        Args:
-            file_type: File type. Support "txt" and "json".
-
-        Return:
-            remote_file_path: Remote file path.
-        """
-        with tempfile.TemporaryDirectory() as tmp_dir:
-            tmp_file_path = self._dump_tmp_file(file_type, tmp_dir)
-
-            # Upload to minio s3 bucket
-            _, remote_file_path = self._remote.upload_file(
-                to_upload=tmp_file_path,
-                project=PROJECT,
-                domain=DOMAIN,
-            )
-
-        return remote_file_path
-
-    def _dump_tmp_file(self, file_type: str, tmp_dir: str) -> str:
-        """Generate and dump a temporary file locally."""
-        if file_type == "txt":
-            tmp_file_path = pathlib.Path(tmp_dir) / "test.txt"
-            with open(tmp_file_path, "w") as f:
-                f.write("Hello World!")
-        elif file_type == "json":
-            d = {"name": "jiawei", "height": 171}
-            tmp_file_path = pathlib.Path(tmp_dir) / "test.json"
-            with open(tmp_file_path, "w") as f:
-                json.dump(d, f)
-
-        return tmp_file_path
-
-    def delete_file(self, bucket: str, key: str) -> None:
-        """Delete the remote file from minio s3 bucket to free the space.
-
-        Args:
-            bucket: s3 bucket name.
-            key: Key name of the object.
-        """
-        res = self._s3_client.delete_object(Bucket=bucket, Key=key)
-        assert res["ResponseMetadata"]["HTTPStatusCode"] == 204
 
 
 @mock.patch.dict(os.environ, {
